requirements.txt
biopython
ete3
pyqt5
